crud lineas 90%

+++ app/Http/Controllers/LineController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\LineResource;
 use App\Models\Brand;
+use App\Models\BrandLine;
 use App\Models\Line;
 use App\Models\Family;
+use App\Models\Product;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\DB;
 
 class LineController extends Controller
 {
@@ -18,7 +20,10 @@ class LineController extends Controller
      */
     public function index()
     {
-        $lines = Line::all();
+        $lines = Line::included()
+                        ->filter()
+                        ->sort()
+                        ->getOrPaginate();
         return LineResource::collection($lines);
     }
 
@@ -56,19 +61,6 @@ class LineController extends Controller
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Line $line)
-    {
-        $families = Family::all();
-        $brands = Brand::all();
-        return view('catalogs.lines.edit', compact('line', 'families', 'brands'));
-    }
-
-    /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
@@ -82,11 +74,15 @@ class LineController extends Controller
             'family_id' => 'required'
         ]);
 
-        $line->update($request->all());
+        $line->update([
+            'cod' => $request->cod,
+            'description' => $request->description,
+            'family_id' => $request->family_id,
+        ]);
 
-        $line->brands()->sync($request->brand_id);
+        // $line->brands()->sync($request->brand_id);
         
-        return redirect()->route('lines.index');
+        return LineResource::make($line);
     }
 
     /**
@@ -97,7 +93,55 @@ class LineController extends Controller
      */
     public function destroy(Line $line)
     {
+        if ($line->brands()->count()) {
+            return response()->json(['error' => 'Esta linea esta siendo usada por alguna marca.'], 405);
+        }
+
         $line->delete();
-        return redirect()->route('lines.index');
+        return LineResource::make($line);
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function addBrands(Request $request, Line $line)
+    {
+        //primero obtener los brand_line_id para la linea seleccionada
+        $brandLines = BrandLine::where('line_id', $line->id)->get();
+        $brandLineIds = [];
+
+        foreach ($brandLines as $brandLine) {
+            $brandLineIds[]= $brandLine['id'];
+        }
+
+        //obtener las marcas que estan en uso
+        $brandsInUse = Product::select('brand_line.brand_id', 'brands.description')
+            ->join('brand_line', 'products.brand_line_id', '=', 'brand_line.id')
+            ->join('brands', 'brand_line.brand_id', '=', 'brands.id')
+            ->whereIn('products.brand_line_id', $brandLineIds)
+            ->groupBy('products.brand_line_id')
+            ->get()->toArray();
+
+        //comparar las marcas en uso con el request de las marcas si falta uno lanzar una excepcion 405 la marca esta en uso
+        foreach ($request->brands as $brand) {
+            foreach ($brandsInUse as $key => $brandInUse) {
+                if ($brand == $brandInUse['brand_id']) {
+                    unset($brandsInUse[$key]);
+                }
+            }
+        }
+
+        // Los elementos que queden el array estan en uso
+        if (count($brandsInUse) > 0) {
+            return response()->json(['error' => 'No puede quitar las marcas.', 'brands' => array_values($brandsInUse)], 422);
+        }
+
+        $line->brands()->sync($request->brands);
+        
+        return LineResource::make($line);
     }
 }
